application working, but if postmans promise is used
