+++ api/promises.js
@@ -1,16 +1,29 @@
 /* eslint-disable import/prefer-default-export */
-const jokeEndpoint = 'https://v2.jokeapi.dev/joke/Programming,Spooky?blacklistFlags=nsfw,religious,political,racist,sexist,explicit&format=txt&type=twopart';
+// const jokeEndpoint = 'https://v2.jokeapi.dev/joke/Programming,Spooky?blacklistFlags=nsfw,religious,political,racist,sexist,explicit&format=txt&type=twopart';
 
+// const jokeRequest = () => new Promise((resolve, reject) => {
+//   fetch(jokeEndpoint, {
+//     method: 'GET',
+//     headers: {
+//       'Content-Type': 'application/json',
+//     },
+//   })
+//     .then((response) => response.json())
+//     .then((data) => console.warn(data))
+//     .catch(reject);
+// });
+
+// export { jokeRequest };
 const jokeRequest = () => new Promise((resolve, reject) => {
-  fetch(jokeEndpoint, {
+  const requestOptions = {
     method: 'GET',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-  })
+    redirect: 'follow'
+  };
+
+  fetch('https://v2.jokeapi.dev/joke/Any?blacklistFlags=nsfw,religious,political,racist,sexist,explicit&type=twopart', requestOptions)
     .then((response) => response.json())
-    .then((data) => resolve(data))
-    .catch(reject);
+    .then((result) => resolve(result))
+    .catch((error) => reject(error));
 });
 
 export { jokeRequest };
